Moved RunStrategy to public api package
