+++ src/main/java/org/concordion/internal/ConcordionBuilder.java
@@ -16,6 +16,7 @@ import org.concordion.api.EvaluatorFactory;
 import org.concordion.api.FailFast;
 import org.concordion.api.FullOGNL;
 import org.concordion.api.Resource;
+import org.concordion.api.RunStrategy;
 import org.concordion.api.Source;
 import org.concordion.api.SpecificationLocator;
 import org.concordion.api.SpecificationReader;
@@ -42,7 +43,6 @@ import org.concordion.internal.command.EchoCommand;
 import org.concordion.internal.command.ExecuteCommand;
 import org.concordion.internal.command.LocalTextDecorator;
 import org.concordion.internal.command.RunCommand;
-import org.concordion.internal.command.RunStrategy;
 import org.concordion.internal.command.SetCommand;
 import org.concordion.internal.command.SpecificationCommand;
 import org.concordion.internal.command.ThrowableCatchingDecorator;
