+++ src/main/java/org/concordion/internal/command/SequentialRunStrategy.java
@@ -3,6 +3,7 @@ package org.concordion.internal.command;
 import org.concordion.api.Resource;
 import org.concordion.api.Result;
 import org.concordion.api.ResultRecorder;
+import org.concordion.api.RunStrategy;
 import org.concordion.api.Runner;
 import org.concordion.internal.FailFastException;
